Fix method signature resolution
When an entire package is missing from the ClassPath, the current binding
resolution fails to identify the parameter's type qualified name,
returning only the first missing package's name, e.g:

> m1/6[pac.B,pac.B,Missing,pac.Missing,Missing2,pacmissing

Note that "pacmissing" is only the package name, the correct would be
"pacmissing.Missing". This PR fixes this, by using the unresolved type
name if the binding type is of the "recovered" type.